chore: move MavenArtifactCopy to common code
