+++ rockcraft/src/main/java/com/canonical/rockcraft/util/MavenArtifactCopy.java
@@ -11,7 +11,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.canonical.rockcraft.gradle;
+package com.canonical.rockcraft.util;
 
 import java.io.File;
 import java.io.IOException;
